+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to QETLAB will be documented in this file.
 
-## Changes since 2015-04-13
+## [0.9] - 2016-01-12
 ### Added
 - BCSGameLB: Computes a lower bound on the quantum value of a binary constraint system (BCS) game.
 - BCSGameValue: Computes the maximum value of a binary constraint system (BCS) game. In classical and no-signalling settings, the value computed is exact, but the quantum value is just an upper bound.
@@ -10,11 +10,14 @@ All notable changes to QETLAB will be documented in this file.
 - L1NormCoherence: Computes the l1-norm of coherence of a quantum state.
 - NonlocalGameLB: Computes a lower bound on the quantum value of a two-player non-local game.
 - RandomPOVM: Computes a random POVM of a specified size and with a specified number of outcomes.
+- RelEntCoherence: Computes the relative entropy of coherence of a quantum state.
 - RobustnessCoherence: Computes the robustness of coherence of a quantum state.
+- TraceDistanceCoherence: Computes the trace distance of coherence of a quantum state.
 - helpers/bcs_to_nonlocal: Converts a description of a binary constraint system (BCS) game into a form that can be presented as a general non-local game.
 - helpers/pure_to_mixed: Converts a state vector or density matrix representation of a state to a density matrix.
 
 ### Changed
+- Entropy: Fixed a bug that would cause NaN output for some low-rank input states.
 - kpNorm: Can now be used as the objective function or as a constraint in a CVX optimization problem, regardless of k and p (only certain special values of k and p were supported previously).
 - kpNormDual: Can now be used as the objective function or as a constraint in a CVX optimization problem, regardless of k and p (only certain special values of k and p were supported previously).
 - NonlocalGameValue: Added the REPT optional input argument, which lets the user specify the number of times that the non-local game will be repeated in parallel.